# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: PawPlanet API
  description: |-
    PawPlanet - Nền tảng mạng xã hội cho thú cưng và người yêu động vật.
    
    Hệ thống bao gồm:
    - **Encyclopedia**: Tra cứu kiến thức về loài vật, giống vật nuôi
    - **Pet Management**: Quản lý hồ sơ thú cưng, tìm người nhận nuôi
    - **Social Network**: Đăng bài, comment, like, theo dõi
    - **User Management**: Quản lý tài khoản, hồ sơ người dùng
    - **Media Upload**: Upload ảnh/video qua Cloudinary

  version: 1.0.0
  contact:
    email: support@pawplanet.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://pawplanet-api.herokuapp.com/api/v1
    description: Production server
...
paths:
...
  /pets:
    get:
      tags:
        - pet
      summary: Lấy danh sách thú cưng
      operationId: getPets
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [owned, for_adoption]
        - name: speciesId
          in: query
          schema:
            type: integer
            format: int64
        - name: breedId
          in: query
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetListResponse'

    post:
      tags:
        - pet
      summary: Tạo hồ sơ thú cưng mới
      operationId: createPet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePetRequest'
      responses:
        '201':
          description: Tạo thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
  /pets/{petId}:
    get:
      tags:
        - pet
      summary: Xem chi tiết thú cưng
      operationId: getPetById
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetDetail'
        '404':
          description: Không tìm thấy pet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - pet
      summary: Cập nhật thông tin thú cưng
      operationId: updatePet
      security:
        - bearerAuth: []
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePetRequest'
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'

    delete:
      tags:
        - pet
      summary: Xóa hồ sơ thú cưng
      operationId: deletePet
      security:
        - bearerAuth: []
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Xóa thành công
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Số trang (bắt đầu từ 0)
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      description: Số lượng items trên 1 trang
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  schemas:
...
    Pet:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        speciesId:
          type: integer
          format: int64
        speciesName:
          type: string
        breedId:
          type: integer
          format: int64
        breedName:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          type: string
          enum: [MALE, FEMALE, UNKNOWN]
        description:
          type: string
        status:
          type: string
          enum: [owned, for_adoption]
        ownerId:
          type: integer
          format: int64
        ownerUsername:
          type: string
        media:
          type: array
          items:
            $ref: '#/components/schemas/PetMedia'
    PetDetail:
      allOf:
        - $ref: '#/components/schemas/Pet'
        - type: object
          properties:
            adoptionProfile:
              $ref: '#/components/schemas/AdoptionProfile'
            followerCount:
              type: integer
            isFollowing:
              type: boolean
              description: User hiện tại có đang follow pet này không
    CreatePetRequest:
      type: object
      required:
        - name
        - speciesId
      properties:
        name:
          type: string
          example: Lucky
        speciesId:
          type: integer
          format: int64
          example: 1
        breedId:
          type: integer
          format: int64
          example: 5
        birthDate:
          type: string
          format: date
          example: "2022-03-15"
        gender:
          type: string
          enum: [MALE, FEMALE, UNKNOWN]
          example: MALE
        description:
          type: string
          example: Chó con rất đáng yêu và năng động
        status:
          type: string
          enum: [owned, for_adoption]
          example: owned
        url:
            type: string
            description: URL của avatar (nếu có)
            example: "https://res.cloudinary.com/pawplanet/image/upload/v1640000000/pawplanet/avatars/user_123.jpg"
#        avatarMediaId:
#          type: integer
#          format: int64
#          description: ID của media avatar (đã upload trước)
#          example: 12
#        mediaIds:
#          type: array
#          items:
#            type: integer
#            format: int64
#          description: Danh sách ID các media trong gallery
#          example: [13, 14, 15]
    UpdatePetRequest:
      type: object
      properties:
        name:
          type: string
        breedId:
          type: integer
          format: int64
        birthDate:
          type: string
          format: date
        gender:
          type: string
        description:
          type: string
        status:
          type: string
    AdoptionProfile:
      type: object
      properties:
        petId:
          type: integer
          format: int64
        healthStatus:
          type: string
        vaccinated:
          type: boolean
        sterilized:
          type: boolean
        personality:
          type: string
        habits:
          type: string
        favoriteActivities:
          type: string
        careInstructions:
          type: string
        diet:
          type: string
        adoptionRequirements:
          type: string
        reasonForAdoption:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PetMedia:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [image, video]
        role:
          type: string
          enum: [avatar, gallery]
        url:
          type: string
        displayOrder:
          type: integer
    PetListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Pet'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer

    # ============ ENCYCLOPEDIA SCHEMAS ============
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: INVALID_INPUT
        message:
          type: string
          example: Email đã tồn tại trong hệ thống
        details:
          type: object
          additionalProperties: true